refactor: extract radiobutton props to separate file
