+++ src/components/index.ts
@@ -8,7 +8,7 @@ export { TextButton, TextButtonProps } from './Button/TextButton';
 export { Input, InputProps } from './Input/Input';
 export { Banner, BannerProps, useBannerDismiss } from './Banner/Banner';
 export { Checkbox, CheckboxProps } from './Checkbox/Checkbox';
-export { RadioButton, RadioButtonProps } from './RadioButton/RadioButton';
+export { RadioButton } from './RadioButton/RadioButton';
 export { Modal, ModalProps, useModalDismiss } from './Modal/Modal';
 export { Textarea, TextAreaProps } from './Textarea/Textarea';
 export { Select, SelectProps } from './Select/Select';
@@ -44,3 +44,4 @@ export { Skeleton, SkeletonProps } from './Skeleton/Skeleton';
 export { Search, SearchProps } from './Search/Search';
 export { Password } from './Password/Password';
 export { VisuallyHidden } from './VisuallyHidden/VisuallyHidden';
+export { RadioButtonProps } from './RadioButton/RadioButtonProps';
